Improve boiler plate code
This plugin originally used the same code to create the Extension object as I have used, only with slight modifications, since I staretd making  extensions. Instead of putting a bunch of vaguley named consdtanst directly in the plugin's namespace the properties of the Extension itself will now be reached through an EXTENSION constant.